Bugfixes for functions with finite domain only
Technically these aren't "finite functions" anymore, but just function
whose *domain* is finite.
Nevertheless, it's very useful to allow for non-finite monoidal signatures.

- Fix bug in `universal`
  - `yarrow.bipartite_muligraph.universal` crashed when applied to diagrams with non-finite signatures
    because the diagram's input dtype was not preserved.
  - Add `test_compose_nonfinite_signature` function.
- More comprehensive tests for finite functions with non-finite codomains
  - Generators for finite-domain functions and composites
  - Tests for coproduct and compositions involving non-finite codomain
- Be more consistent in setting default dtype in yarrow.array.numpy
- When FiniteFunction is passed a backend array, don't try and create a new one;
  just use it directly.
- Check table dtypes when taking coproducts of finite functions

+++ yarrow/bipartite_multigraph.py
@@ -237,7 +237,7 @@ def universal(q: AbstractFiniteFunction, f: AbstractFiniteFunction):
     such that q ; u = f.
     """
     target = f.target
-    u = q._Array.zeros(q.target)
+    u = q._Array.zeros(q.target, dtype=f.table.dtype)
     # TODO: in the below we assume the PRAM CRCW model: multiple writes to the
     # same memory location in the 'u' array can happen in parallel, with an
     # arbitrary write succeeding.
